refactor(BoardCanvas): extract the logic for reordering lists and cards into standalone functions
